Added misc camera setters

+++ Emu/include/Tiles/TileMap.h
@@ -17,8 +17,8 @@ namespace Engine
 		EMU_API TileMap(const std::string mapFile, const int numMetersPerTile);
 		EMU_API const char GetTile(int x, int y) const;
 		
-		EMU_API inline size_t GetWidth() const { return m_width; }
-		EMU_API inline size_t GetHeight() const { return m_height; }
+		EMU_API inline int GetWidth() const { return m_width; }
+		EMU_API inline int GetHeight() const { return m_height; }
 
 	public:
 		inline std::vector<Tile>& GetCollisionBodies() { return m_collisionBodies; }
@@ -37,8 +37,8 @@ namespace Engine
 		std::vector<char> m_map;
 		std::vector<Tile> m_collisionBodies;
 		std::vector<Tile> m_tiles;
-		size_t m_width;
-		size_t m_height;
+		int m_width;
+		int m_height;
 
 		int m_numMetersPerTile;
 	};
